+++ src/Support/helpers/functions.php
@@ -60,3 +60,38 @@ if ( ! function_exists('logOnLocal')) {
         }
     }
 }
+
+
+if ( ! function_exists('getArraySubset')) {
+    /**
+     * Returns the given <keys> in <array>. Returns a formatted string,
+     * a single value, or an array with the values for the keys.
+     * If any of the <keys> is empty and $ignoreEmpty == true
+     * that record will be filtered from the results
+     *
+     * @param array $array
+     * @param string|array $keys
+     * @param string $format
+     * @param bool $ignoreEmpty
+     *
+     * @return array
+     */
+    function getArraySubset(array $array, string|array $keys, string $format = "", bool $ignoreEmpty = true): array
+    {
+        $parsedElements = array_map(function ($row) use ($keys) {
+
+            // If any of the keys is empty return null
+            $values = [];
+            foreach (( is_array($keys) ? $keys : [$keys]) as $k){
+                if( empty($row[$k]) )
+                    return null;
+
+                $values[] = $row[$k];
+            }
+
+            return empty($format) ? ( is_array($keys) ? $values : $values[0] ) : vsprintf($format, $values);
+        }, $array);
+
+        return $ignoreEmpty ? array_filter($parsedElements) : $parsedElements;
+    }
+}